rename: notification -> notifications 라우팅 명칭 변경, 적용
