+++ pages/myinfo/constants/myInfoLinks.ts
@@ -1,0 +1,14 @@
+import { ROUTES } from '@/constants/routes';
+
+export const MYINFO_NOTIFICATIONS = [
+  {
+    key: 'notification',
+    name: '알림',
+    pathname: ROUTES.MY_INFO.NOTIFICATON,
+  },
+  {
+    key: 'subscribe',
+    name: '구독한 기업',
+    pathname: ROUTES.MY_INFO.SUBSCRIBE,
+  },
+];
